Worked On Level 1
I placed blocks, which act as the ground that the player will walk on. Previously, there was a small problem in which the player could not walk, and would just fall through the ground. To fix this problem, I placed a box collider 2D instead of a regular box collider. This had to be done because the player was 2D. Next time I will make sure to pay attention which colliders I need to place.