Worked On Player Teleportation To Level 2
There was one major issue. The player was not teleporting to the next scene (Level 2). To solve this issue, I tried changing the size of the portal, played around with colliders, and also tried to work on code for the key. Next time I will try to make sure the player can enter the portal, only once they collected the key.