Almost Done With Level 3
I am almost done with placing everything in Level 3. I wanted to add arrows to the scene, but they originally had a white background, instead of a transparent one. However, I imported them as transparent images. So, I simply checked some certain things, applied it, and placed it in the scene. It was then able to work. Next time, I will place arrows all over my game to guide the player, showing them which direction they should be going to.