Final Version
In this game, players must look for a key around the map to activate a portal. The maps are different, but they also come with various difficulty levels. Some levels are harder than others, while some are much easier. In order to complete the game, players must complete five levels.